requirements.txt
Django==2.2.9
django-taggit==1.2.0
django-widget-tweaks==1.4.5
Markdown==3.2.1
Pillow==7.0.0
psycopg2==2.8.4
pytz==2019.3
sqlparse==0.3.0
